display_all_teacher_data

+++ src/module/TeacherALL/Component/TeacherALL.tsx
@@ -1,18 +1,22 @@
 import React from "react"
 import "../style/TeacherALL.style.css"
+import "../style/RespoTable.style.css"
 import apiServicesAllTeacher from "../../API/AllTeacher.API"
 import { useNavigate } from "react-router-dom"
+import { typeResponzeALLTechers } from "../../API/types.API"
+import { typeResponzeALLTechersObject } from "../../API/types.API"
+
 
 function TeacherALL(): JSX.Element {
-const location = useNavigate
+    const location = useNavigate()
+    const [AllReachersResponzeDATA, setAllReachersResponzeDATA] = React.useState<typeResponzeALLTechers>()
     React.useEffect(() => {
         const JWTToken = localStorage.getItem("authenticationKey") as string
         apiServicesAllTeacher.apiAllTeacher(JWTToken)
-            .then((data: any) => {
-               /*  if (!data) {
+            .then((data: typeResponzeALLTechers) => {
+                if (!data) {
                     localStorage.clear(); location("LoginPage")
-                    
-                } */
+                } else { setAllReachersResponzeDATA(data) }
             })
             .catch(err => console.log(err))
     }, [])
@@ -23,7 +27,17 @@ const location = useNavigate
                 <h1>All list Teacher</h1>
             </div>
             <div className="teacherALLContent">
-                zgxvszgxvzsvxz
+                {
+                    AllReachersResponzeDATA?.map((item: typeResponzeALLTechersObject, key: number) => (
+                        <div 
+                        key={key}
+                        className="respoDATA">
+                            <div className="ResId"><h1>{item.id}</h1></div>
+                            <div className="Resname"><h1>{item.name}</h1></div>
+                            <div className="Ressubject"><h1>{item.subject}</h1></div>
+                        </div>
+                    ))
+                }
             </div>
         </div>
     )
